refactor(meilisearch): [Issue #40] optimisation and file placement

+++ src/Persistence/Playbook.Persistence.Meilisearch/SearchLogger.cs
@@ -2,7 +2,7 @@
 
 public partial class SearchLogger(ILogger<SearchLogger> logger)
 {
-    public void LogSearchPerformance(string? term, long elapsedMs, int count)
+    public void LogSearchPerformance(string? term, long elapsedMs, long count)
     {
         if (!logger.IsEnabled(LogLevel.Information)) return;
 
@@ -13,5 +13,5 @@ public partial class SearchLogger(ILogger<SearchLogger> logger)
         EventId = 1001,
         Level = LogLevel.Information,
         Message = "Search executed for '{Term}' in {ElapsedMs}ms. Results: {Count}")]
-    static partial void LogSearchPerformanceCore(ILogger logger, string? term, long elapsedMs, int count);
+    static partial void LogSearchPerformanceCore(ILogger logger, string? term, long elapsedMs, long count);
 }
